Fixed KituraTemplateEngine import error in Xcode
When using Xcode to build a Kitura-based project, compiling fails because the module `KituraTemplateEngine` can't be found.

The problem comes from the dash in the template engine's name. Xcode automatically changes `-` to `_`. Using the `Xcode` compiler macro, that is present in all projects generated using the SPM command line tool, to reflect this change and provide the correct module name in Xcode fixes the compiling issue, without breaking existing source or impacting command line builds.