Store URL encoded key-value pairs as a [String: [String]] rather than a [String: String].
When there are multiple values for a key, append a new value to the array rather than concatcenating the values with a comma in a string. This will simplify counting how many values there are for a key and avoid issues where a value itself might have a comma, but may cause breakage with existing code.