Cache URLParser objects
URLParser objects are created once in every RouterRequest. Their creation involves a conversion of String representation of a URL to Data followed by a parsing step. Typically, a webservice consists of a limited number of routes, and hence a limited number of target URLs. It may hence make sense to cache the URLParser objects against the String representations of the URLs. Since we are talking about a limited number of URLs here, the possibility of a leak is negligible. Caching can lead to a significant performance gain in certain workloads.