Add a StaticFileServer option to fallback to the index page
Single page applications use only one index file that is accessible by web
browsers, which is usually the index.html, which also serves as a landing page.
All the other pages are accessible only from the landing page. If the users
bypasses the index page to directly visit an inner page, or if they try to
refresh an inner page, the static file server returns a 404 error. Some SPAs
might wish to have a fallback feature where the static file server
serves the index.html in the event of such user activity.

This commit adds a new option to the StaticFileServer.Options class to enable
fallback to the index.html. This is intended to be used only by single page
applications. Just before the static file server returns a 404, we check if
this option is configured and serve the default index page if it is. The webapp
takes care of the rendering.